Extract Drag and Drop methods
